+++ number/mod.cpp
@@ -1,7 +1,7 @@
 #include "../template.cpp"
 
-long long powm(long long a, long long n, long long m) {
-    long long ret = 1;
+int64_t powm(int64_t a, int64_t n, int64_t m) {
+    int64_t ret = 1;
     while (n > 0) {
         if (n & 1) (ret *= a) %= m;
         (a *= a) %= m;
@@ -10,6 +10,6 @@ long long powm(long long a, long long n, long long m) {
     return ret;
 }
 
-long long invm(long long a, long long m) {
+int64_t invm(int64_t a, int64_t m) {
     return powm(a, m-2, m);
 }
